Complete conversion to RVVM Machine API
- Rewrote most of the stuff to use RVVM Machine API instead of riscv32_
- Removed riscv32_ legacy
- Implemented hart-local decoder to dynamically switch ISA
- Implemented A extension with atomic lib, lr/sc on top of CAS
- Better CLINT/PLIC interrupt routines using riscv_interrupt
- New arguments parser, mem/smp/rv64 options

- Multi-core VMs now work excellent
- Dhrystone reports 15% more performance, RAM speed is improved by 30%

- Fixed rwfunc arguments mismatch in riscv_mmio_scan
- Fixed dev->end comparsion in riscv_mmio_scan
- Fixed riscv_fetch_inst & riscv_mmu_fetch_inst on page-misaligned operations
- Fixed amomin/amomax atomic operations (typo)
- Fixed riscv_trap with no SJLJ by using riscv_restart_dispatch
- Fixed vector iterators scope
- Fixed empty mmio machine pointer
- Fixed EthOC DMA corruption on read < 0
- Fixed warnings about ringbuf warn format on mingw

+++ src/ringbuf.c
@@ -46,7 +46,7 @@ bool ringbuf_is_empty(struct ringbuf *rb)
 bool ringbuf_put(struct ringbuf *rb, void *data, size_t len)
 {
 	if (ringbuf_get_free_spc(rb) < len) {
-		rvvm_warn("Overflow in ringbuf %p! size=%zu, consumed=%zu, len=%zu", rb, rb->size, rb->consumed, len);
+		rvvm_warn("Overflow in ringbuf %p! size=%u, consumed=%u, len=%u", rb, (uint32_t)rb->size, (uint32_t)rb->consumed, (uint32_t)len);
 		return false;
 	}
